Add support for Python 3
The majority of changes needed to support Python 3 are using print
functions and absolute imports.

There are also changes to the code generation, as in multiple places
unordered collections (e.g. sets, dicts) are used to keep track of
variables, and these are evaluated in different orders between Python 2
and Python 3. Occurences of these have been changed to ordered
collections using sorted() and OrderedDict().